Split local and remote middleware functions out into own files along with splitting the tests out
