+++ core/middleware/middleware.go
@@ -1,14 +1,11 @@
 package middleware
 
 import (
-	"bytes"
-	"encoding/json"
 	"os"
 	"os/exec"
 	"path"
 	"strings"
 
-	"errors"
 	"io/ioutil"
 	"net/http"
 
@@ -140,131 +137,14 @@ func (this *Middleware) Execute(pair models.RequestResponsePair) (models.Request
 	}
 }
 
-// ExecuteMiddleware - takes command (middleware string) and payload, which is passed to middleware
-func (this Middleware) executeMiddlewareLocally(pair models.RequestResponsePair) (models.RequestResponsePair, error) {
-	commandAndArgs := []string{this.Binary, this.Script.Name()}
-
-	middlewareCommand := exec.Command(commandAndArgs[0], commandAndArgs[1:]...)
-
-	// getting payload
-	pairViewBytes, err := json.Marshal(pair.ConvertToRequestResponsePairView())
-
-	if err != nil {
-		return pair, errors.New("Failed to marshal request to JSON")
-	}
-
-	log.WithFields(log.Fields{
-		"middleware": this.toString(),
-		"stdin":      string(pairViewBytes),
-	}).Debug("preparing to modify payload")
-
-	var stdout bytes.Buffer
-	var stderr bytes.Buffer
-
-	// Redirect standard streams
-	middlewareCommand.Stdin = bytes.NewReader(pairViewBytes)
-	middlewareCommand.Stdout = &stdout
-	middlewareCommand.Stderr = &stderr
-
-	if err := middlewareCommand.Start(); err != nil {
-		log.WithFields(log.Fields{
-			"sdtdout": string(stdout.Bytes()),
-			"sdtderr": string(stderr.Bytes()),
-			"error":   err.Error(),
-		}).Error("Middleware failed to start")
-		return pair, err
-	}
-
-	if err := middlewareCommand.Wait(); err != nil {
-		log.WithFields(log.Fields{
-			"sdtdout": string(stdout.Bytes()),
-			"sdtderr": string(stderr.Bytes()),
-			"error":   err.Error(),
-		}).Error("Middleware failed to stop successfully")
-		return pair, err
-	}
-
-	// log stderr, middleware executed successfully
-	if len(stderr.Bytes()) > 0 {
-		log.WithFields(log.Fields{
-			"sdtderr": string(stderr.Bytes()),
-		}).Info("Information from middleware")
-	}
-
-	if len(stdout.Bytes()) > 0 {
-		var newPairView RequestResponsePairView
-
-		err = json.Unmarshal(stdout.Bytes(), &newPairView)
-
-		if err != nil {
-			return pair, errors.New("Failed to unmarshal JSON from middleware")
-		} else {
-			if log.GetLevel() == log.DebugLevel {
-				log.WithFields(log.Fields{
-					"middleware": this.toString(),
-					"payload":    string(stdout.Bytes()),
-				}).Debug("payload after modifications")
-			}
-			// payload unmarshalled into RequestResponsePair struct, returning it
-			return models.NewRequestResponsePairFromRequestResponsePairView(newPairView), nil
+func buildLogFields(fields map[string]string) log.Fields {
+	logFields := log.Fields{}
+	for field, _ := range fields {
+		if fields[field] != "" {
+			logFields[field] = fields[field]
 		}
-	} else {
-		log.WithFields(log.Fields{
-			"stdout": string(stdout.Bytes()),
-		}).Warn("No response from middleware.")
 	}
-
-	return pair, nil
-
-}
-
-func (this Middleware) executeMiddlewareRemotely(pair models.RequestResponsePair) (models.RequestResponsePair, error) {
-	pairViewBytes, err := json.Marshal(pair.ConvertToRequestResponsePairView())
-
-	if this.Remote == "" {
-		return pair, fmt.Errorf("Error when communicating with remote middleware")
-	}
-
-	req, err := http.NewRequest("POST", this.Remote, bytes.NewBuffer(pairViewBytes))
-	req.Header.Add("Content-Type", "application/json")
-	if err != nil {
-		log.WithFields(log.Fields{
-			"error": err.Error(),
-		}).Error("Error when building request to remote middleware")
-		return pair, err
-	}
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"error": err.Error(),
-		}).Error("Error when communicating with remote middleware")
-		return pair, err
-	}
-
-	if resp.StatusCode != 200 {
-		log.Error("Remote middleware did not process payload")
-		return pair, errors.New("Error when communicating with remote middleware")
-	}
-
-	returnedPairViewBytes, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"error": err.Error(),
-		}).Error("Error when process response from remote middleware")
-		return pair, err
-	}
-
-	var newPairView RequestResponsePairView
-
-	err = json.Unmarshal(returnedPairViewBytes, &newPairView)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"error": err.Error(),
-		}).Error("Error when trying to serialize response from remote middleware")
-		return pair, err
-	}
-	return models.NewRequestResponsePairFromRequestResponsePairView(newPairView), nil
+	return logFields
 }
 
 func (this Middleware) IsSet() bool {
